better checkmate detection, reduction in permutations, turn hints
this is a good change, i tested the speed and its a bit better anywhere from 3-5 times faster to only a bit better. the only time its worse is when there is one legal move, which is not very common, and even then it is acceptable. I really need to test this though